app/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
packaging==21.3
pandas==1.4.3
patsy==0.5.2
psycopg2-binary==2.9.3
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
scipy==1.9.0
six==1.16.0
statsmodels==0.13.2
urllib3==1.26.11
Werkzeug==2.2.2
zipp==3.8.1
